+++ packages/ui/jest.config.cjs
@@ -1,0 +1,11 @@
+/** @type {import('jest').Config} */
+module.exports = {
+  testEnvironment: 'jsdom',
+  clearMocks: true,
+  setupFilesAfterEnv: ['<rootDir>/jest.setup.cjs'],
+  testMatch: ['**/?(*.)+(spec|test).js'],
+  transform: {
+    '^.+\\.jsx?$': 'babel-jest'
+  },
+  transformIgnorePatterns: ['node_modules/(?!theme-ui)']
+}